Remove duplicate uuid to bytearray helper.
